"South-east" map data.

Created in /data-raw/ semap is the sptable version of some of maptools 'wrld_simpl' and seatt is the matching attribute data, linked by 'object_'.

Created in /data-raw/.

Examples

# recreate as sp object
mp <- sp(semap, attr_tab = seatt, crs = "+proj=longlat +ellps=WGS84")
#> New names:
#> • `object_` -> `object_...1`
#> • `object_` -> `object_...7`