Convert between different types of spatial objects.

Facilities for converting between different types of spatial objects, including an in-place method to modify the underlying geometry of 'Spatial' classes using data frame idioms.The spbabel package provides functions to round-trip a Spatial object to a single table and back.

Details

sptable<-

modify a Spatial object in-place

sptable

create a tibble from Spatial DataFrame object

sp

create Spatial DataFrame object from table